third_party/blink/web_tests/virtual/prefetch-sw/external/wpt/speculation-rules/prefetch/tentative/service-worker/basic.sub.https_origin=same-site_sw=race-fetch-handler-modify-url-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch with ServiceWorker (race-fetch-handler-modify-url)
  assert_equals: Race network request should be sent to the server with Sec-Purpose. expected 1 but got 0
Harness: the test ran to completion.
